+++ src/algorithm/mod.rs
@@ -38,7 +38,7 @@ fn max_shape(a: &[usize], b: &[usize]) -> Shape {
 pub trait FillContext: Copy {
     type Error;
     fn error(self, msg: impl ToString) -> Self::Error;
-    fn pierce_boxes(self) -> bool;
+    fn tip_boxes(self) -> bool;
     fn fill<T: ArrayValue>(self) -> Option<T>;
     fn fill_error(error: Self::Error) -> Self::Error;
     fn is_fill_error(error: &Self::Error) -> bool;
@@ -49,8 +49,8 @@ impl FillContext for &Uiua {
     fn error(self, msg: impl ToString) -> Self::Error {
         self.error(msg)
     }
-    fn pierce_boxes(self) -> bool {
-        self.pierce_boxes()
+    fn tip_boxes(self) -> bool {
+        self.tip_boxes()
     }
     fn fill<T: ArrayValue>(self) -> Option<T> {
         T::get_fill(self)
@@ -68,8 +68,8 @@ impl FillContext for &&mut Uiua {
     fn error(self, msg: impl ToString) -> Self::Error {
         (**self).error(msg)
     }
-    fn pierce_boxes(self) -> bool {
-        (**self).pierce_boxes()
+    fn tip_boxes(self) -> bool {
+        (**self).tip_boxes()
     }
     fn fill<T: ArrayValue>(self) -> Option<T> {
         T::get_fill(self)
@@ -87,7 +87,7 @@ impl FillContext for () {
     fn error(self, msg: impl ToString) -> Self::Error {
         panic!("{}", msg.to_string())
     }
-    fn pierce_boxes(self) -> bool {
+    fn tip_boxes(self) -> bool {
         false
     }
     fn fill<T: ArrayValue>(self) -> Option<T> {